+++ .project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>v1.2</name>
+	<name>Group-Project</name>
 	<comment></comment>
 	<projects>
 	</projects>
