--- WordGames/ProjectA/src/com/GuessWordGame.java
@@ -1,10 +1,0 @@
-package com;
-
-public class GuessWordGame {
-	
-	public static void main(String[] args)
-	{
-		new GuessWordFrame();
-	}
-
-}
