--- WordGames/ProjectWordCounter/WordCountGame.java
@@ -1,11 +1,0 @@
-package com;
-
-public class WordCountGame {
-	
-	public static void main(String[] args)
-	{
-		//creates the object of user interface for the WOrd Count Game
-		new WordCountFrame();
-	}
-
-}
