--- WordGames/ProjectWordCounter/src/com/WordCountGame.java
@@ -1,10 +1,0 @@
-package com;
-
-public class WordCountGame {
-	
-	public static void main(String[] args)
-	{
-		new WordCountFrame();
-	}
-
-}
