+++ src/com/ProjectCountWord/WordCountGame.java
@@ -1,0 +1,8 @@
+package com.ProjectCountWord;
+
+public class WordCountGame {
+	
+	public static void main(String[] args) {
+		new WordCountFrame();
+	}
+}
