+++ src/com/ProjectGuessWord/GuessWordGame.java
@@ -1,0 +1,8 @@
+package com.ProjectGuessWord;
+
+public class GuessWordGame {
+	
+	public static void main(String[] args) {
+		new GuessWordFrame();
+	}
+}
